Implemented better integer inference. Also improved float accuracy to double accuracy

+++ app/Main.hs
@@ -17,7 +17,7 @@ data Opts = Opts {
   language :: Language,
   verbosity :: Int,
   countBranches :: Bool,
-  topKCutoff :: Maybe Float
+  topKCutoff :: Maybe Double
 } deriving Show
 
 readLanguage :: ReadM Language
